+++ src/rules/selector-list-comma-newline-after/index.js
@@ -32,7 +32,7 @@ export default function (expectation) {
       // a, /* comment */
       // b {}
       const selector = (rule.raws.selector) ? rule.raws.selector.raw : rule.selector
-      styleSearch({ source: selector, target: "," }, match => {
+      styleSearch({ source: selector, target: ",", outsideFunctionalNotation: true }, match => {
         const nextThreeChars = selector.substr(match.endIndex, 3)
 
         // If there's a // comment, that means there has to be a newline